Part4/Chapter02-02-wordGame/requirements.txt
scikit-learn==1.6.1
streamlit==1.45.1
openai==1.79.0
python-dotenv==1.1.0